Split diracx.db into diracx.db.sql and diracx.db.os
