+++ setup.cfg
@@ -74,12 +74,14 @@ console_scripts =
     dirac = diracx.cli:app
 diracx =
 	properties_module = diracx.core:properties
-diracx.dbs =
+diracx.db.sql =
 	AuthDB = diracx.db:AuthDB
 	JobDB = diracx.db:JobDB
 	JobLoggingDB = diracx.db:JobLoggingDB
 	SandboxMetadataDB = diracx.db:SandboxMetadataDB
 	#DummyDB = diracx.db:DummyDB
+diracx.db.os =
+	JobParametersDB = diracx.db:JobParametersDB
 diracx.services =
 	jobs = diracx.routers.job_manager:router
 	config = diracx.routers.configuration:router
