+++ src/diracx/db/__init__.py
@@ -1,7 +1,7 @@
 __all__ = ("AuthDB", "JobDB", "JobLoggingDB", "SandboxMetadataDB")
 
-from .auth.db import AuthDB
-from .jobs.db import JobDB, JobLoggingDB
-from .sandbox_metadata.db import SandboxMetadataDB
+from .sql.auth.db import AuthDB
+from .sql.jobs.db import JobDB, JobLoggingDB
+from .sql.sandbox_metadata.db import SandboxMetadataDB
 
 # from .dummy.db import DummyDB
