+++ src/diracx/db/os/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+__all__ = ("JobParametersDB",)
+
+from .job_parameters import JobParametersDB
