+++ src/diracx/db/sql/dummy/db.py
@@ -4,7 +4,8 @@ from uuid import UUID
 
 from sqlalchemy import func, insert, select
 
-from ..sql_utils import BaseSQLDB, apply_search_filters
+from diracx.db.sql.utils import BaseSQLDB, apply_search_filters
+
 from .schema import Base as DummyDBBase
 from .schema import Cars, Owners
