+++ src/diracx/db/sql/jobs/status_utility.py
@@ -9,7 +9,7 @@ from diracx.core.models import (
     SetJobStatusReturn,
 )
 from diracx.core.utils import JobStatus
-from diracx.db.jobs.db import JobDB, JobLoggingDB
+from diracx.db.sql.jobs.db import JobDB, JobLoggingDB
 
 
 async def set_job_status(
