+++ src/diracx/db/sql/utils.py
@@ -82,7 +82,9 @@ class BaseSQLDB(metaclass=ABCMeta):
         """Return the available implementations of the DB in reverse priority order."""
         db_classes: list[type[BaseSQLDB]] = [
             entry_point.load()
-            for entry_point in select_from_extension(group="diracx.dbs", name=db_name)
+            for entry_point in select_from_extension(
+                group="diracx.db.sql", name=db_name
+            )
         ]
         if not db_classes:
             raise NotImplementedError(f"Could not find any matches for {db_name=}")
@@ -96,8 +98,7 @@ class BaseSQLDB(metaclass=ABCMeta):
         prefixed with ``DIRACX_DB_URL_{DB_NAME}``.
         """
         db_urls: dict[str, str] = {}
-        # TODO: rename diracx.dbs to diracx.sql_dbs
-        for entry_point in select_from_extension(group="diracx.dbs"):
+        for entry_point in select_from_extension(group="diracx.db.sql"):
             db_name = entry_point.name
             var_name = f"DIRACX_DB_URL_{entry_point.name.upper()}"
             if var_name in os.environ:
