+++ tests/db/auth/test_authorization_flow.py
@@ -4,7 +4,7 @@ import pytest
 from sqlalchemy.exc import NoResultFound
 
 from diracx.core.exceptions import AuthorizationError
-from diracx.db.auth.db import AuthDB
+from diracx.db.sql.auth.db import AuthDB
 
 MAX_VALIDITY = 2
 EXPIRED = 0
