+++ tests/db/auth/test_refresh_token.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import pytest
 
-from diracx.db.auth.db import AuthDB
-from diracx.db.auth.schema import RefreshTokenStatus
+from diracx.db.sql.auth.db import AuthDB
+from diracx.db.sql.auth.schema import RefreshTokenStatus
 
 
 @pytest.fixture
